Modify message counter.
I applied the change to the message counter. Now only show the number one the place and number, establish a correlation between number picture place and number picture number. 0 with 0, 1 with 1, ... and show de correct place and number to establish an association. Because for an unknown reason I cant find the error in how to ordered properly the pictures. Show in place number 12 the association between 100 and 100.